Extending/correcting stimulation helper
Extending and correcting logic for the stimulation helper. Also moved status checks to directly below Notes.

I would like to make the results display cleaner (merge the four separate paragraphs) and possibly create an inline 'stimulation ardor check die' similar to the 'valiant moxie check die', but I mostly do not know how to work with the formatting. 

It might also be nice if we could make sections collapsible as the screen is getting kind of busy with my new additional sections.